Failed to add support to PLY models
TinyPly does not support lists with different number of elements per entry, meaning, you cannot tell how many vertices has a face. And there is no way to verify they are all the same so I will read wrong triangles in some cases silently. So it is dissabled for now until I find it useful to implement it properly.