Fixes on BVH construction
Fixed a problem where some triangles were skipped.
This happened with "thin" triangles, where one axis was significant smaller than the other two.

Improved BVH construction.
Split buckets computed as divisions of the centroids space instead of the AABB space.
Split is tested on the tree axis for each node. The one producing a split closer to 50/50 is the one selected.
This also fixes another problem with thin triangles as the previous method could stop splitting as a split in the larger axis put all the triangles in the same bucket.
This method could be up to 50% slower at creation time but raycasting is faster because of the better subdivision.

Added some code for easy logging, like task timing. In VS can be shown in the debug output window.