BVH generation improvements
Fully implemented the paper "On fast Construction of SAH-based Bounding Volume Hierarchies".

Our code checks all three axis for each subdivision but the least cost axis is not always the fastests for our raycast. In one of our two test meshes using the least cost axis pro is about 40% slower when calculating AO. But in the other mesh using the largest axis is 400% slower. So sticking to the least cost axis for now. Need more test meshes though.