Fix mesh mapping problem
Some models still presented problems with thin triangles.
No amount of padding on the barycentric tests were able to remove the problems completely.
Moved barycentric calculations to double precision.
Changed the code to make sure that the barycentric coordinates are computed only if necessary to make the double precision calculations to hit performance as little as possible.