New mesh mapping method
Now mesh mapping can be done using the low-poly mesh normals or with a smooth version of the mesh.
The new method (smooth mesh) produces better results for non continuous normals on different sides of an edge. But it also produces skewing in the middle of the faces.
A hybrid method must come to solve this (and probably add new problems)