Hybrid mesh mapping + out of range uvs
Adding a not very useful (for now) hybrid method to map low mesh to high mesh.
The idea is to allow hard edges on the low poly mesh and compute a proper transition between faces while keeping other areas untouched.

Also, fixed a crash when the low-poly mesh has any UVs out of the 0 to 1 range.